requirements.txt
api==0.0.7
Django==5.1.1
django_rest_passwordreset==1.4.1
djangorestframework==3.15.2
Requests==2.32.3
rest_framework_simplejwt==0.0.2
